+++ home-care-center-start/src/main/java/edu/hust/controller/BedController.java
@@ -5,8 +5,11 @@ import edu.hust.common.util.RandomUUID;
 import edu.hust.common.vo.ApiResult;
 import edu.hust.dao.dto.Bed;
 import edu.hust.monitor.Monitor;
+import edu.hust.service.domain.ClientFull;
 import edu.hust.service.service.BedService;
 import edu.hust.common.exception.GlobalException;
+import edu.hust.service.service.ClientService;
+import edu.hust.service.service.RoomService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,8 +29,14 @@ import java.util.List;
 @RequestMapping("HomeCareCenter/bed/")
 public class BedController {
 
-    @Resource
+    @Autowired
     private BedService bedService;
+
+    @Autowired
+    private RoomService roomService;
+
+    @Autowired
+    private ClientService clientService;
 
     @Autowired
     private RandomUUID randomUUID;
@@ -61,6 +70,9 @@ public class BedController {
     @PostMapping("add")
     @Monitor("addBed")
     public ApiResult add(@RequestBody Bed bed) {
+        if (!legal(bed)) {
+            throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+        }
         bed.setId(randomUUID.nextIdStr());
         bedService.addBed(bed);
         return ApiResult.buildSuccess();
@@ -71,6 +83,9 @@ public class BedController {
     @Monitor("addBatchBed")
     public ApiResult addBatch(@RequestBody List<Bed> bedList) {
         for (Bed bed : bedList) {
+            if (!legal(bed)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+            }
             bed.setId(randomUUID.nextIdStr());
         }
         bedService.addBedList(bedList);
@@ -81,6 +96,13 @@ public class BedController {
     @PostMapping("update")
     @Monitor("updateBed")
     public ApiResult update(@RequestBody Bed bed) {
+        if (
+                bed.getRoom() != null
+                && bed.getRoom().getId() != null
+                && roomService.getRoomById(bed.getRoom().getId()) == null
+        ) {
+            throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+        }
         bedService.updateBed(bed);
         return ApiResult.buildSuccess();
     }
@@ -93,16 +115,49 @@ public class BedController {
             @RequestParam(value = "roomId", required = false) String roomId
     ) {
         if (id == null && roomId == null) {
+            List<ClientFull> clientFullList = clientService.getClientInfoList();
+            if (!clientFullList.isEmpty()) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             bedService.deleteAllBed();
             return ApiResult.buildSuccess();
         }else if (id == null) {
+            List<Bed> bedList = bedService.getBedByRoomId(roomId);
+            for (Bed bed : bedList) {
+                if (!legalDelete(bed.getId())) {
+                    throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+                }
+            }
             bedService.deleteBedByRoomId(roomId);
             return ApiResult.buildSuccess();
         }else if (roomId == null) {
+            if (!legalDelete(id)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             bedService.deleteBedById(id);
             return ApiResult.buildSuccess();
         }else {
             throw new GlobalException(ApiCodeEnum.PARAM_ERROR);
         }
+    }
+
+    private boolean legal(Bed bed) {
+        if (
+                bed.getBedTitle() == null
+                || bed.getRoom() == null
+                || bed.getRoom().getId() == null
+                || roomService.getRoomById(bed.getRoom().getId()) == null
+        ) {
+            return false;
+        }
+        return true;
+    }
+
+    private boolean legalDelete(String bedId) {
+        ClientFull clientFull = clientService.getClientInfoByBedId(bedId);
+        if (clientFull == null) {
+            return true;
+        }
+        return false;
     }
 }
