+++ home-care-center-start/src/main/java/edu/hust/controller/DishController.java
@@ -4,15 +4,20 @@ import edu.hust.common.constant.ApiCodeEnum;
 import edu.hust.common.util.RandomUUID;
 import edu.hust.common.vo.ApiResult;
 import edu.hust.dao.dto.Dish;
+import edu.hust.dao.dto.DishSet;
 import edu.hust.monitor.Monitor;
+import edu.hust.service.domain.DishSetFull;
 import edu.hust.service.service.DishService;
 import edu.hust.common.exception.GlobalException;
+import edu.hust.service.service.DishSetService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 
 /**
@@ -26,8 +31,11 @@ import java.util.List;
 @RequestMapping("HomeCareCenter/dish/")
 public class DishController {
 
-    @Resource
+    @Autowired
     private DishService dishService;
+
+    @Autowired
+    private DishSetService dishSetService;
 
     @Autowired
     private RandomUUID randomUUID;
@@ -57,6 +65,9 @@ public class DishController {
     @PostMapping("add")
     @Monitor("addDish")
     public ApiResult add(@RequestBody Dish dish) {
+        if (!legal(dish)) {
+            throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+        }
         dish.setId(randomUUID.nextIdStr());
         dishService.addDish(dish);
         return ApiResult.buildSuccess();
@@ -67,6 +78,9 @@ public class DishController {
     @Monitor("addBatchDish")
     public ApiResult addBatch(@RequestBody List<Dish> dishList) {
         for (Dish dish : dishList) {
+            if (!legal(dish)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+            }
             dish.setId(randomUUID.nextIdStr());
         }
         dishService.addDishList(dishList);
@@ -89,16 +103,58 @@ public class DishController {
             @RequestParam(value = "name", required = false) String name
     ) {
         if (id == null && name == null) {
+            List<DishSetFull> dishSetFullList = dishSetService.getDishSetList();
+            if (!dishSetFullList.isEmpty()) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             dishService.deleteAllDish();
             return ApiResult.buildSuccess();
         } else if (id == null) {
+            Dish dish = dishService.getDishByName(name);
+            if (dish == null) {
+                return ApiResult.buildSuccess();
+            }
+            if (!legalDelete(dish.getId())) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             dishService.deleteDishByName(name);
             return ApiResult.buildSuccess();
         } else if (name == null) {
+            if (!legalDelete(id)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             dishService.deleteDishById(id);
             return ApiResult.buildSuccess();
         } else {
             throw new GlobalException(ApiCodeEnum.PARAM_ERROR);
         }
+    }
+
+    private boolean legal(Dish dish) {
+        if (dish.getName() == null) {
+            return false;
+        }
+        return true;
+    }
+
+    private boolean legalDelete(String dishId) {
+        List<DishSetFull> dishSetFullList = dishSetService.getDishSetList();
+        for (DishSetFull dishSetFull : dishSetFullList) {
+            List<String> dishList = new ArrayList<>();
+            dishList.add(dishSetFull.getMon());
+            dishList.add(dishSetFull.getTue());
+            dishList.add(dishSetFull.getWed());
+            dishList.add(dishSetFull.getThu());
+            dishList.add(dishSetFull.getFri());
+            dishList.add(dishSetFull.getSat());
+            dishList.add(dishSetFull.getSun());
+            String ids = String.join(",", dishList);
+            String str[] = ids.split(",");
+            List<String> idList = Arrays.asList(str);
+            if (idList.contains(dishId)) {
+                return false;
+            }
+        }
+        return true;
     }
 }
