+++ home-care-center-start/src/main/java/edu/hust/controller/RoomController.java
@@ -3,8 +3,11 @@ package edu.hust.controller;
 import edu.hust.common.constant.ApiCodeEnum;
 import edu.hust.common.util.RandomUUID;
 import edu.hust.common.vo.ApiResult;
+import edu.hust.dao.dto.Bed;
 import edu.hust.dao.dto.Room;
 import edu.hust.monitor.Monitor;
+import edu.hust.service.service.AreaService;
+import edu.hust.service.service.BedService;
 import edu.hust.service.service.RoomService;
 import edu.hust.common.exception.GlobalException;
 import io.swagger.annotations.Api;
@@ -26,8 +29,14 @@ import java.util.List;
 @RequestMapping("HomeCareCenter/room/")
 public class RoomController {
 
-    @Resource
+    @Autowired
     private RoomService roomService;
+
+    @Autowired
+    private AreaService areaService;
+
+    @Autowired
+    private BedService bedService;
 
     @Autowired
     private RandomUUID randomUUID;
@@ -61,6 +70,9 @@ public class RoomController {
     @PostMapping("add")
     @Monitor("addRoom")
     public ApiResult add(@RequestBody Room room) {
+        if (!legal(room)) {
+            throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+        }
         room.setId(randomUUID.nextIdStr());
         roomService.addRoom(room);
         return ApiResult.buildSuccess();
@@ -71,6 +83,9 @@ public class RoomController {
     @Monitor("addBatchRoom")
     public ApiResult addBatch(@RequestBody List<Room> roomList) {
         for (Room room : roomList) {
+            if (!legal(room)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+            }
             room.setId(randomUUID.nextIdStr());
         }
         roomService.addRoomList(roomList);
@@ -81,6 +96,13 @@ public class RoomController {
     @PostMapping("update")
     @Monitor("updateRoom")
     public ApiResult update(@RequestBody Room room) {
+        if (
+                room.getArea() != null
+                && room.getArea().getId() != null
+                && areaService.getAreaInfoById(room.getArea().getId()) == null
+        ) {
+            throw new GlobalException(ApiCodeEnum.ILLEGAL_DATA);
+        }
         roomService.updateRoom(room);
         return ApiResult.buildSuccess();
     }
@@ -93,12 +115,25 @@ public class RoomController {
             @RequestParam(value = "areaId", required = false) String areaId
     ) {
         if (id == null && areaId == null) {
+            List<Bed> bedList = bedService.getBedList();
+            if (!bedList.isEmpty()) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             roomService.deleteAllRoom();
             return ApiResult.buildSuccess();
         } else if (id == null) {
+            List<Room> roomList = roomService.getRoomByAreaId(areaId);
+            for (Room room : roomList) {
+                if (!legalDelete(room.getId())) {
+                    throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+                }
+            }
             roomService.deleteRoomByAreaId(areaId);
             return ApiResult.buildSuccess();
         } else if (areaId == null) {
+            if (!legalDelete(id)) {
+                throw new GlobalException(ApiCodeEnum.ILLEGAL_DELETE);
+            }
             roomService.deleteRoomById(id);
             return ApiResult.buildSuccess();
         } else {
@@ -106,5 +141,24 @@ public class RoomController {
         }
     }
 
+    private boolean legal(Room room) {
+        if (
+                room.getRoomTitle() == null
+                || room.getArea() == null
+                || room.getArea().getId() == null
+                || areaService.getAreaInfoById(room.getArea().getId()) == null
+        ) {
+            return false;
+        }
+        return true;
+    }
+
+    private boolean legalDelete(String roomId) {
+        List<Bed> bedList = bedService.getBedByRoomId(roomId);
+        if (bedList.isEmpty()) {
+            return true;
+        }
+        return false;
+    }
 }
